Refactored components/GrandmasterMode
