+++ src/components/Game.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Grid } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import GameClock from '../features/gameClock/GameClock';
-import StartedButtonsGrandmasterMode from './GrandmasterMode/StartedButtonsGrandmasterMode';
+import StartedButtonsGrandmasterMode from '../features/grandmasterMode/StartedButtonsGrandmasterMode';
 import FinishedButtonsPlayMode from './PlayMode/FinishedButtonsPlayMode';
 import FinishedDialogsPlayMode from './PlayMode/FinishedDialogsPlayMode';
 import StartedButtonsPlayMode from './PlayMode/StartedButtonsPlayMode';
